backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.9.2
websockets==12.0
